requirements/main.txt
einops
librosa
local_attention
matplotlib
numpy
onnx
onnxruntime
# onnxruntime-gpu   # for gpu
# onnxsim
PyYAML
pyworld
resampy
scikit_learn
scipy
SoundFile
tensorboard
torchcrepe
torchfcpe
tqdm
transformers
